old template

+++ rest-webservice/src/main/java/com/vchizhande/webservice/dto/RentalDto.java
@@ -12,7 +12,7 @@ public class RentalDto {
 
     private Long id;
 
-    private Integer currentMileage;
+    private Long currentMileage;
 
     private double penaltyFee;
